refactor(lib): Put all the code inside a "lib" subdirectory.
