+++ lib/bans/index.js
