+++ lib/bans/sqs.js
