+++ lib/config/index.js
