+++ test/local/email_record_tests.js
@@ -3,7 +3,7 @@
 
 require('ass')
 var test = require('tap').test
-var emailRecord = require('../../email_record')
+var emailRecord = require('../../lib/email_record')
 
 function now() {
   return 1000 // old school
